backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
pyyaml==6.0.1
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.26.2
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
SQLAlchemy==2.0.23
aiosqlite==0.19.0
